+++ BackEnd/TacoHub/src/main/java/com/example/TacoHub/Entity/NotionCopyEntity/PageEntity.java
@@ -2,6 +2,8 @@ package com.example.TacoHub.Entity.NotionCopyEntity;
 
 import com.example.TacoHub.Entity.BaseDateEntity;
 import jakarta.persistence.*;
+import lombok.Builder;
+import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.core.annotation.Order;
@@ -13,6 +15,7 @@ import java.util.UUID;
 @Entity
 @Getter
 @Setter
+@Builder
 @Table(name = "page") // Specify the table name if needed
 public class PageEntity extends BaseDateEntity {