+++ BackEnd/TacoHub/src/main/java/com/example/TacoHub/Entity/NotionCopyEntity/WorkSpaceEntity.java
@@ -2,6 +2,7 @@ package com.example.TacoHub.Entity.NotionCopyEntity;
 
 import com.example.TacoHub.Entity.BaseDateEntity;
 import jakarta.persistence.*;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.web.service.annotation.GetExchange;
@@ -13,6 +14,7 @@ import java.util.UUID;
 @Entity
 @Getter
 @Setter
+@Builder
 @Table(name = "workspace") // Specify the table name if needed
 public class WorkSpaceEntity extends BaseDateEntity {
     @Id
@@ -24,6 +26,6 @@ public class WorkSpaceEntity extends BaseDateEntity {
     private String name; // Workspace Name
 
     @OneToMany(mappedBy = "workspace", fetch = FetchType.LAZY, cascade = CascadeType.ALL)
-    private List<PageEntity> pages = new ArrayList<>();
+    private List<PageEntity> rootPages = new ArrayList<>();
 
 }