+++ BackEnd/TacoHub/src/main/java/com/example/TacoHub/Repository/NotionCopyRepository/PageRepository.java
@@ -1,4 +1,11 @@
 package com.example.TacoHub.Repository.NotionCopyRepository;
 
-public interface PageRepository {
+import com.example.TacoHub.Entity.NotionCopyEntity.PageEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.UUID;
+
+@Repository
+public interface PageRepository extends JpaRepository<PageEntity, UUID> {
 }