+++ BackEnd/TacoHub/src/main/java/com/example/TacoHub/Repository/NotionCopyRepository/WorkSpaceUserRepository.java
@@ -1,11 +1,12 @@
 package com.example.TacoHub.Repository.NotionCopyRepository;
 
-import com.example.TacoHub.Entity.NotionCopyEntity.WorkSpaceEntity;
+import com.example.TacoHub.Entity.NotionCopyEntity.WorkSpaceUserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.UUID;
 
 @Repository
-public interface WorkSpaceRepository extends JpaRepository<WorkSpaceEntity, UUID> {
+public interface WorkSpaceUserRepository extends JpaRepository<WorkSpaceUserEntity, UUID> {
+
 }